refactor: 파일 분리 및 clsx로 교체
