+++ package.json
@@ -18,6 +18,7 @@
     "@radix-ui/react-slot": "^1.1.0",
     "@vanilla-extract/css": "^1.15.5",
     "@vanilla-extract/recipes": "^0.5.5",
+    "clsx": "^2.1.1",
     "jotai": "^2.9.3",
     "lottie-react": "^2.4.0",
     "next": "14.2.8",
